fix(vllm_performance): k8s resource not marked for cleaning on exception (#230)
fix(vllm_perf): k8s resource not marked for cleaning on exception

If there was an exception in port_foward creation or benchmark:
- the port-forward was not cleaned up
- done_using is not called on the environment

This is because the clean-up code required an successfully benchmark to execute and was not in a finally: block.

The impact of this would be the inability of the actuator to run an experiment (continued attempts to run same entity that cannot succeed) if
* max_environments environments had been created
* there was at least one exception among the entities processed by each environment
* an entity needed an new environment.

Neither of the existing environments would be "reaped" as they entities with the exceptions never decremented the in use counter (so it could never reach zero).

The entity requiring the new environment would find all "max_environment"