fix(custom_experiments): detect if custom experiment returns unexpected properties (#250)
fix(custom_experiments): Flag if custom experiment function outputs do not match the defined outputs

The custom_experiment decorator defines output names.

These names must be keys in the dict returned by the function.

There was no check if when executing the function if the names matched - this would lead to a error which was handled but the error description did not describe the source of the issue.